- Added a new abstract class that simplifies further xml serialization testing - Refactored other classes accordingly
