--- src/test/resources/partial_examples/invoice_body/document_general/pension_funds/DatiCassaPrevidenziale_3.xml
@@ -1,27 +1,0 @@
-<DatiGeneraliDocumento>
-    <TipoDocumento>TD01</TipoDocumento>
-    <Divisa>EUR</Divisa>
-    <Data>2012-03-10</Data>
-    <Numero>A-2012-15</Numero>
-    <DatiRitenuta>
-        <TipoRitenuta>RT01</TipoRitenuta>
-        <ImportoRitenuta>240.25</ImportoRitenuta>
-        <AliquotaRitenuta>20.00</AliquotaRitenuta>
-        <CausalePagamento>A</CausalePagamento>
-    </DatiRitenuta>
-    <DatiCassaPrevidenziale>
-        <TipoCassa>TC22</TipoCassa>
-        <AlCassa>4.00</AlCassa>
-        <ImportoContributoCassa>60.00</ImportoContributoCassa>
-        <ImponibileCassa>1500.00</ImponibileCassa>
-        <AliquotaIVA>21.00</AliquotaIVA>
-        <Ritenuta>SI</Ritenuta>
-    </DatiCassaPrevidenziale>
-    <DatiCassaPrevidenziale>
-        <TipoCassa>TC03</TipoCassa>
-        <AlCassa>2.00</AlCassa>
-        <ImportoContributoCassa>31.20</ImportoContributoCassa>
-        <ImponibileCassa>1560.00</ImponibileCassa>
-        <AliquotaIVA>21.00</AliquotaIVA>
-    </DatiCassaPrevidenziale>
-</DatiGeneraliDocumento>
